Tests for PSBT
Summary:
Added functional tests for PSBT that test the RPCs. Also added all
of the BIP 174 test vectors (except for the updater tests) in the
functional tests.

Added a Unit test for the BIP 174 updater test vector.

This is a partial backport of Core PR13557 : https://github.com/bitcoin/bitcoin/pull/13557/commits/020628e3a4e88e36647eaf92bac4b3552796ac6a

Depends on D4352

Test Plan:
  make check
  ./test/functionnal/test_runner.py rpc_psbt

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D4353

+++ src/wallet/rpcwallet.cpp
@@ -5180,7 +5180,7 @@ bool FillPSBT(const CWallet *pwallet, PartiallySignedTransaction &psbtx,
 
     // Fill in the bip32 keypaths and redeemscripts for the outputs so that
     // hardware wallets can identify change
-    for (unsigned int i = 0; i < txConst->vout.size(); ++i) {
+    for (size_t i = 0; i < txConst->vout.size(); ++i) {
         const CTxOut &out = txConst->vout.at(i);
         PSBTOutput &psbt_out = psbtx.outputs.at(i);
